New parser: planet available buildings parser. many changes:
XNPlanetBuildingItem.__str__() now will show seconds left/total if build is in progress
XNPlanet: chaged the way it stores info about available buildings and buildings in progress. Changed the way to add info about build in progress and detect if building is in progress

world: first, parse planet available buildings and store them inside planet structure. Then scan for builds in progress on the same page, and add build in progress to planet, if any.